# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_pos-retail_go_backend_internal_types_request.CreatePurchaseOrderRequest:
    properties:
      expected_date:
        type: string
      items:
        items:
          $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_request.PurchaseOrderItemRequest'
        type: array
      notes:
        type: string
      supplier_id:
        type: string
      warehouse_id:
        type: string
    type: object
  github_com_pos-retail_go_backend_internal_types_request.CreateReturnRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_request.ReturnItemRequest'
        type: array
      reason:
        type: string
      refund_method:
        type: string
      sale_id:
        type: string
      warehouse_id:
        type: string
    type: object
...
  github_com_pos-retail_go_backend_internal_types_request.PurchaseOrderItemRequest:
    properties:
      discount:
        type: number
      product_id:
        type: string
      quantity:
        type: integer
      tax_rate:
        type: number
      unit_price:
        type: number
    type: object
...
  github_com_pos-retail_go_backend_internal_types_request.ReturnItemRequest:
    properties:
      condition:
        type: string
      notes:
        type: string
      product_id:
        type: string
      quantity:
        type: integer
      sale_item_id:
        type: string
    type: object
...
  github_com_pos-retail_go_backend_internal_types_request.UpdatePurchaseOrderStatusRequest:
    properties:
      status:
        type: string
    type: object
...
  github_com_pos-retail_go_backend_internal_types_response.ApiResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  github_com_pos-retail_go_backend_internal_types_response.PaginatedResponse:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.Pagination'
      success:
        type: boolean
    type: object
  github_com_pos-retail_go_backend_internal_types_response.Pagination:
    properties:
      has_more:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
info:
  contact: {}
  description: Go (Fiber) implementation of POS Retail Backend.
  title: POS Retail Backend API
  version: "1.0"
paths:
...
  /api/purchases:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: PO status
        in: query
        name: status
        type: string
      - description: Supplier ID
        in: query
        name: supplier_id
        type: string
      - description: Warehouse ID
        in: query
        name: warehouse_id
        type: string
      - default: 50
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: List purchase orders
      tags:
      - Purchases
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Purchase order payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_request.CreatePurchaseOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Create purchase order
      tags:
      - Purchases
  /api/purchases/{id}:
    delete:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Purchase Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Cancel purchase order
      tags:
      - Purchases
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Purchase Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get purchase order
      tags:
      - Purchases
  /api/purchases/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Purchase Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Status payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_request.UpdatePurchaseOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Update purchase order status
      tags:
      - Purchases
  /api/returns:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Warehouse ID
        in: query
        name: warehouse_id
        type: string
      - description: Sale ID
        in: query
        name: sale_id
        type: string
      - description: Return status
        in: query
        name: status
        type: string
      - default: 50
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: List returns
      tags:
      - Returns
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Return payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_request.CreateReturnRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Create return
      tags:
      - Returns
  /api/returns/{id}:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Return ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_pos-retail_go_backend_internal_types_response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get return
      tags:
      - Returns
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"